+++ src/TeleFlow.Telegram.Client/Properties/AssemblyInfo.cs
@@ -1,3 +1,6 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("TeleFlow.Framework")]
+[assembly: InternalsVisibleTo("TeleFlow.Telegram.LongPolling")]
+[assembly: InternalsVisibleTo("TeleFlow.Telegram.Webhooks")]
+[assembly: InternalsVisibleTo("TeleFlow.ArchitectureTests")]